backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
google-genai>=1.0.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform
matplotlib>=3.8.0
python-pptx>=0.6.23
pillow
requests
python-dotenv
httpx
aiofiles
